Module Bechamel__Analyze

Analyze module.

Micro-benchmark usually uses a linear-regression to estimates the execution time of a code segments. For example, the following table might represent {!Measurement_raw.t} array collected by Benchmark.run:

  +-----+------+
  | run | time |
  +-----+------+
  | 1   | 19   |
  | 2   | 25   |
  | 3   | 37   |
  | 4   | 47   |
  | 5   | 56   |
  +-----+------+

Bechamel records 3000 samples and the number of iterations can grows geometrically (see Benchmark.run). Then, Bechamel can use 2 algorithms:

Ordinary Least Square

RANdom SAmple Consensus

The user can choose one of it. Currently, OLS is the best to use. These algorithms will estimate the actual execution time of the code segment. Using OLS with the above data would yield an estimated execution time of 9.6 nanoseconds with a goodness of fit (r²) of 0.992.

More generally, Bechamel lets the user to choose predictors and the responder. Indeed, the user can use others metrics (such as perf) and the API allows to analyze such metrics each other.